+++ organize/__init__.py
@@ -2,11 +2,10 @@ import logging
 import logging.config
 import os
 
-import appdirs
-import colorama
+import appdirs  # type: ignore
+import colorama  # type: ignore
 import yaml
 
-from .__version__ import __version__
 from .compat import Path
 
 colorama.init(autoreset=True)
@@ -16,7 +15,7 @@ APP_DIRS = appdirs.AppDirs("organize")
 
 # setting the $ORGANIZE_CONFIG env variable overrides the default config path
 if os.getenv("ORGANIZE_CONFIG"):
-    CONFIG_PATH = Path(os.getenv("ORGANIZE_CONFIG")).resolve()
+    CONFIG_PATH = Path(os.getenv("ORGANIZE_CONFIG", "")).resolve()
     CONFIG_DIR = CONFIG_PATH.parent
 else:
     CONFIG_DIR = Path(APP_DIRS.user_config_dir)
@@ -55,6 +54,8 @@ handlers:
 root:
     level: DEBUG
     handlers: [file]
+exifread:
+    level: INFO
 """.format(
     filename=str(LOG_PATH)
 )