+++ tests/conftest.py
@@ -1,9 +1,16 @@
 import os
+from typing import Iterable, Tuple, Union
+from unittest.mock import patch
 
 import pytest
 from mock import patch
 
 from organize.compat import Path
+from organize.utils import DotDict
+
+TESTS_FOLDER = os.path.dirname(os.path.abspath(__file__))
+
+TESTS_FOLDER = os.path.dirname(os.path.abspath(__file__))
 
 
 def create_filesystem(tmp_path, files, config):
@@ -13,7 +20,7 @@ def create_filesystem(tmp_path, files, config):
             name, content = f
         except Exception:
             name = f
-            content = ''
+            content = ""
         p = tmp_path / "files" / Path(name)
         p.parent.mkdir(parents=True, exist_ok=True)
         with p.open("w") as ptr: